--- src/main/java/blackjack/domain/PlayingCards.java
@@ -1,23 +1,0 @@
-package blackjack.domain;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-public class PlayingCards {
-    private final List<PlayingCard> playingCards = new ArrayList<>();
-
-    public void addCard(PlayingCard playingCard) {
-        playingCards.add(playingCard);
-    }
-
-    public List<PlayingCard> getPlayingCards() {
-        return Collections.unmodifiableList(playingCards);
-    }
-
-    public int getResult() {
-        return playingCards.stream()
-            .mapToInt(playingCard -> playingCard.getDenomination().getScore())
-            .sum();
-    }
-}
