+++ src/test/java/blackjack/domain/PlayerTest.java
@@ -3,6 +3,9 @@ package blackjack.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import blackjack.domain.card.Denomination;
+import blackjack.domain.card.PlayingCard;
+import blackjack.domain.card.Suit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -102,8 +105,8 @@ class PlayerTest {
 
         // then
         assertAll(
-            () -> assertThat(isBurstForDealer).isTrue(),
-            () -> assertThat(isBurstForGambler).isTrue()
+                () -> assertThat(isBurstForDealer).isTrue(),
+                () -> assertThat(isBurstForGambler).isTrue()
         );
     }
 }
