+++ src/test/java/blackjack/domain/PlayingCardTest.java
@@ -1,8 +1,11 @@
 package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
+import blackjack.domain.card.Denomination;
+import blackjack.domain.card.PlayingCard;
+import blackjack.domain.card.Suit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -21,8 +24,8 @@ class PlayingCardTest {
 
         // then
         assertAll(
-            () -> assertThat(suit).isEqualTo(expectedSuit),
-            () -> assertThat(denomination).isEqualTo(expectedDenomination)
+                () -> assertThat(suit).isEqualTo(expectedSuit),
+                () -> assertThat(denomination).isEqualTo(expectedDenomination)
         );
     }
 }
